requirements.txt
asgiref==3.7.2
black==24.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.2.10
djangorestframework==3.14.0
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
